Support Payment Response over SSL
This adds the abililty to set whether WorldPay should use an SSL (https)
connection to the payment response page. The benifit of this is that the
callback password will never be sent in the clear. WorldPay does not seem
to care if the site is using a self signed certificate so this is a very
useful feature for sites that do not have an SSL certificate signed by the
certificate authorities.

Because user's are likley to use this in combination with a site not using
SSL for general purpose viewing, another option has been added to complement
it: "Force non-ssl links". This option allows the theme to decided whether or
not links back to the site should avoid HTTPS. The reason for this is that
the shopper response pages are generated from the payment response URL at the
HTTPS address which leads Drupal to append https:// to all absolute links. If
a user was to click such a link with https on a self-signed cert site they
would recieve the usual scary browser message which would likley lead to a lot
of customers screaming "scam".