fixContact and hideContact now work
The options to prevent editing of and hidding of contact details now works.
WorldPay's main documentation sugest that they simply need to exist and do not
need a value:
http://www.worldpay.com/support/kb/bg/htmlredirect/rhtml5205.html
But another pice documentation (which I needed google to find) states that the
value must be set to "true":
http://www.worldpay.com/support/kb/bg/customisingadvanced/custa9105.html

Keep your bloomin documentation upto date and consistant WorldPay.