Auto redirect payment page
Switched on the auto redirect so the user does not need to click the button
anymore.

I've added a bit of a hack for debuging. If the debug is set to screen or both
then auto-redirect is disabled. Kinda not useful to show the data being posted
to the screen if the browser throws you to another page.