Implement a custom hook
Now other modules can append additional parameters to the WorldPay post. For
exmaple. I am going to be using this to add the user's phone number.

It could also be useful for theme developers as they can create custom WorldPay
parameters (C_var and MC_var) that can be used in theme substitutions.