Issue #1538050 by jozzhart, magicmyth: Fix WorldPay transaction timestamp
WorldPay returns a timestamp in the JAVA format which means it is in
miliseconds. It seems depending on the the system, this even converted
to an incorrect int value (because the number was too long) or caused an SQL
failure. This solves both issues by dividing the value by 1000 which will
implicitly cast it to a float.

Thanks go to jozzhart for catching this bug.