New $site_id variable for brand adjustments
A company may have different sides to its brand but as a single company only
have one WorldPay account. This makes it difficult to represent the specific
brand in use on the payment pages. Using the new Site ID setting the brand
can be target by constructing image and CSS class names using this
variable in a <WPDISPLAY> tag like so:

<img src="<WPDISPLAY ITEM=C_siteId-ppe empty='fallback_id'>-logo.png" />